icfp: fix not-so-good example
The example is trying to illustrate something about the first-order
embedding, but uses a program with the same behavior on first-order
and higher-order

Change to a MUCH more complicated program that's different with FO and
HO.

A simpler program would use mutable data ... but that's out of the
question. It's hard to think of anything else (simpler!) where an
untyped function can dis-respect its argument.